overtime, manifest, user, softdelete, too much

+++ app/Http/Controllers/OptionsController.php
@@ -6,6 +6,7 @@ use App\Models\Division;
 use App\Models\JobTitle;
 use App\Models\Position;
 use App\Models\Department;
+use App\Models\Employee;
 use Illuminate\Http\Request;
 use App\Models\EmployeeStatus;
 use App\Models\Holiday;
@@ -71,6 +72,15 @@ class OptionsController extends Controller
 
 //position delete
     public function positionDelete($id){
+
+        $exist = Employee::where('position_id', $id)->exists();
+        if ($exist) {
+            return response()->json([
+                'success' => false,
+                'message' => 'This position cannot be deleted because it is assigned to an employee.',
+            ]);
+        }
+
         $position = Position::find($id);
         $position->delete();
 
@@ -79,7 +89,6 @@ class OptionsController extends Controller
             'message' => 'The position has been deleted.',
             'redirect' => route('options.list') 
         ]);
-        // return redirect()->route('options.list')->with('success', 'Position deleted successfully.');
     }
         
 //jobtitle add
@@ -113,6 +122,14 @@ class OptionsController extends Controller
 
 //jobtitle delete
     public function jobTitleDelete($id){
+        $exist = Employee::where('job_title_id', $id)->exists();
+        if ($exist) {
+            return response()->json([
+                'success' => false,
+                'message' => 'This job title cannot be deleted because it is assigned to an employee.',
+            ]);
+        }
+
         $jobTitle = JobTitle::find($id);
         $jobTitle->delete();
         
@@ -148,8 +165,16 @@ class OptionsController extends Controller
     }
 
 //division delete
-    public function divisionDelete($id)
-    {
+    public function divisionDelete($id){
+
+        $exist = Employee::where('division_id', $id)->exists();
+        if ($exist) {
+            return response()->json([
+                'success' => false,
+                'message' => 'This division cannot be deleted because it is assigned to an employee.',
+            ]);
+        }
+        
         $division = Division::find($id);
         $division->delete();
         
@@ -191,6 +216,14 @@ class OptionsController extends Controller
 
 //department delete
     public function departmentDelete($id){
+        $exist = Employee::where('department_id', $id)->exists();
+        if ($exist) {
+            return response()->json([
+                'success' => false,
+                'message' => 'This department    cannot be deleted because it is assigned to an employee.',
+            ]);
+        }
+
         $department = Department::find($id);
         $department->delete();
         
@@ -228,6 +261,14 @@ class OptionsController extends Controller
 
 //Employee Status delete
     public function statusDelete($id){
+        $exist = Employee::where('employee_status', $id)->exists();
+        if ($exist) {
+            return response()->json([
+                'success' => false,
+                'message' => 'This employee status cannot be deleted because it is assigned to an employee.',
+            ]);
+        }
+
         $status = EmployeeStatus::findOrFail($id);
         $status->delete();
         
@@ -280,6 +321,15 @@ class OptionsController extends Controller
 //Office Location Delete
     public function deleteLocation($id){
         $ol = OfficeLocation::find($id);
+        $exist = $ol->employees()->exists();
+        if ($exist) {
+            return response()->json([
+                'success' => false,
+                'message' => 'This location cannot be deleted because it is assigned to an employee.',
+            ]);
+        }
+
+        
         $ol->delete();
         
         return response()->json([
